:bug: Fix: 건의 및 버그 제보 제거 시 미확인 상태 여부 검증 로직 추가

+++ src/main/java/kr/modusplant/domains/member/adapter/controller/MemberAdminController.java
@@ -76,6 +76,9 @@ public class MemberAdminController {
     public void removeProposalOrBug(ProposalOrBugReportRemoveRecord record) {
         ReportId reportId = ReportId.create(record.reportUlid());
         if (reportRepository.isIdExistInProposalOrBugReport(reportId)) {
+            if (reportRepository.isUncheckedInProposalOrBugReport(reportId)) {
+                throw new ExistsValueException(NOT_FOUND_PROPOSAL_OR_BUG_REPORT_CHECKED, "checkedAt");
+            }
             reportRepository.removeProposalOrBugReport(reportId);
         }
     }
